% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToTitleCase.R
\name{ToTitleCase}
\alias{ToTitleCase}
\title{Convert a single word to title case}
\usage{
ToTitleCase(x)
}
\arguments{
\item{x}{A single word in the form of a string.}
}
\value{
The original string, but with the first character capitalized.
}
\description{
Convert a single word to title case
}
\examples{
ToTitleCase('quebec')
}
\author{
Patrick Barks <patrick.barks@gmail.com>
}
